PlaceSection 컴포넌트 도입 및 API 통합 완료. Closes #5
