+++ tasty-client/src/api/index.js
@@ -9,3 +9,13 @@ export const getPlaces = async () => {
     return [];
   }
 };
+
+export const getSavedPlaces = async () => {
+  try {
+    const res = await axios.get("http://localhost:3000/users/places");
+    return res.data;
+  } catch (err) {
+    console.error("API 호출 에러:", err);
+    return [];
+  }
+};
